@@ -3,7 +3,7 @@ description: Heyo messeging application
 version: 1.0.0+1
 publish_to: none
 environment:
-  sdk: ">=3.1.2 <4.0.0"
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
   firebase_core: ^2.21.0
@@ -21,7 +21,7 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_svg: ^2.0.7
-
+  package_info_plus: ^5.0.1
   flutter_screenutil: ^5.8.4
   smooth_page_indicator: ^1.0.0+2
   pin_code_fields: ^7.4.0
@@ -42,7 +42,6 @@ dependencies:
   #     url: git@ssh.git.energy:heyo/flutter_camera
   #     ref: development
   #
-  photo_manager: ^2.7.1
   mime: ^1.0.2
   chewie: ^1.3.3
   just_audio: ^0.9.24
@@ -81,7 +80,7 @@ dependencies:
   path_provider: ^2.0.11
   open_file: ^3.2.1
   pull_to_refresh: ^2.0.0
-
+  photo_manager: ^3.0.0-dev.5
   photo_view: ^0.14.0
   gallery_saver: ^2.3.2
   flutter_ringtone_player: ^3.2.0
@@ -109,7 +108,9 @@ dependencies:
   freezed_annotation: ^2.4.1
   json_annotation: ^4.8.1
   flutter_animate: ^4.4.0
-
+dependency_overrides:
+  device_info_plus: 9.1.2
+  win32: ^5.2.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
