@@ -2,12 +2,14 @@ import 'dart:async';
 
 import 'package:get/get.dart';
 import 'package:heyo/app/modules/messages/connection/wifi_direct_connection_provider.dart';
+import 'package:heyo/app/modules/messages/views/messages_view.dart';
 import 'package:heyo/app/modules/shared/data/models/messaging_participant_model.dart';
 import 'package:heyo/app/modules/new_chat/data/models/user_model/user_model.dart';
 import 'package:heyo_wifi_direct/heyo_wifi_direct.dart';
 
 import '../../../routes/app_pages.dart';
 import '../../messages/connection/wifi_direct_connection_controller.dart';
+import '../../messages/utils/chat_Id_generator.dart';
 import '../../shared/data/models/messages_view_arguments_model.dart';
 import '../../wifi_direct/controllers/wifi_direct_wrapper.dart';
 
@@ -17,7 +19,10 @@ class WifiDirectConnectController extends GetxController {
   late WifiDirectConnectionProvider wifiDirectConnectionProvider;
   late HeyoWifiDirect _pluginInstance;
   late StreamSubscription _eventListener;
-  late UserModel user;
+
+  late MessagesViewArgumentsModel args;
+  late String coreId;
+  late String chatId;
 
   WifiDirectConnectController() {
     _pluginInstance = WifiDirectWrapper.pluginInstance!;
@@ -31,12 +36,14 @@ class WifiDirectConnectController extends GetxController {
 
     _eventListener = WifiDirectWrapper.pluginInstance!.consumerEventSource.stream
         .listen((event) => eventHandler(event));
-    user = Get.arguments as UserModel;
+    args = Get.arguments as MessagesViewArgumentsModel;
+    coreId = args.participants.first.coreId;
+    chatId = args.participants.first.chatId;
   }
 
   Future<bool> startConnection() async {
-    if (_pluginInstance.peerList.contains(user.coreId)) {
-      await _pluginInstance.connectPeer(user.coreId);
+    if (_pluginInstance.peerList.contains(coreId)) {
+      await _pluginInstance.connectPeer(coreId);
     } else {
       return false;
     }
@@ -52,8 +59,8 @@ class WifiDirectConnectController extends GetxController {
       case EventType.peerListRefresh:
 
         // PeerList peerList = signaling.wifiDirectPlugin.peerList;
-        if (_pluginInstance.peerList.contains(user.coreId)) {
-          connectionStatus.value = (_pluginInstance.peerList.peers[user.coreId]!).status;
+        if (_pluginInstance.peerList.contains(coreId)) {
+          connectionStatus.value = (_pluginInstance.peerList.peers[coreId]!).status;
         } else {
           connectionStatus.value = PeerStatus.peerUnavailable;
         }
@@ -67,7 +74,7 @@ class WifiDirectConnectController extends GetxController {
         // connectedPeer = event.message as Peer;
         //   wifiDirectConnectionController.eventHandler(event);
         connectionStatus.value =
-            (_pluginInstance.peerList.peers[user.coreId] as Peer).status as PeerStatus;
+            (_pluginInstance.peerList.peers[coreId] as Peer).status as PeerStatus;
 
         print('WifiDirectConnectController: linked to ${(event.message as Peer).multiAddress}');
 
@@ -90,7 +97,7 @@ class WifiDirectConnectController extends GetxController {
 
     switch (connectionStatus.value) {
       case PeerStatus.peerConnected:
-        print('WifiDirectConnectView peer Connected ${user.coreId}');
+        print('WifiDirectConnectView peer Connected ${coreId}');
 
         Future.delayed(const Duration(milliseconds: 500), () {
           Get
@@ -101,8 +108,8 @@ class WifiDirectConnectController extends GetxController {
                 connectionType: MessagingConnectionType.wifiDirect,
                 participants: [
                   MessagingParticipantModel(
-                    coreId: user.coreId,
-                    chatId: user.coreId,
+                    coreId: coreId,
+                    chatId: chatId,
                   ),
                 ],
               ),
@@ -111,7 +118,7 @@ class WifiDirectConnectController extends GetxController {
         break;
 
       case PeerStatus.peerTCPOpened:
-        print('WifiDirectConnectView peer TCPOpened ${user.coreId}');
+        print('WifiDirectConnectView peer TCPOpened ${coreId}');
 
         Future.delayed(const Duration(milliseconds: 500), () {
           Get
@@ -122,8 +129,8 @@ class WifiDirectConnectController extends GetxController {
                 connectionType: MessagingConnectionType.wifiDirect,
                 participants: [
                   MessagingParticipantModel(
-                    coreId: user.coreId,
-                    chatId: user.coreId,
+                    coreId: coreId,
+                    chatId: chatId,
                   ),
                 ],
               ),
