@@ -22,10 +22,7 @@ class CallHistoryDetailListTileWidget extends StatelessWidget {
       onTap: () {
         //* Close keyboard before opening bottom sheet
         FocusScope.of(context).unfocus();
-        Get.find<UserPreview>().openUserPreview(
-          coreId: coreId,
-          name: name,
-        );
+        Get.find<UserPreviewController>().openUserPreview(coreId: coreId);
       },
       //dense: true,
       visualDensity: const VisualDensity(vertical: -4),
