@@ -63,11 +63,11 @@ class ContactListWithHeader extends GetView<NewChatController> {
                   onTap: () {
                     controller.inputFocusNode.unfocus();
                     var contact = contacts[index];
-                    Get.find<UserPreview>().openUserPreview(
+                    Get.find<UserPreviewController>().openUserPreview(
                       coreId: contact.coreId,
-                      name: contact.name,
-                      // isVerified: contact.isVerified,
-                      // isContact: contact.isContact,
+                      //name: contact.name,
+                      //isVerified: contact.isVerified,
+                      //isContact: contact.isContact,
                     );
                   },
                   child: UserWidget(
